test(agent): build the flock and kill probes only on unix
The lease probe (syscall.Flock and its lock constants) and the process
liveness probe (syscall.Kill) do not compile on Windows, and a runtime
GOOS skip does not stop the build. The helpers and every test that needs
them now live in //go:build unix files, in both agent and execenv.
