+++ agent/execenv/local_env_test.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 	"os"
 	"path/filepath"
-	"runtime"
 	"strings"
 	"sync"
 	"testing"
@@ -292,34 +291,6 @@ func TestUnsandboxedScratchDirConcurrentProvisioning(t *testing.T) {
 		if got != results[0] {
 			t.Fatalf("goroutine %d got scratch dir %q, want %q (every caller must share one dir)", i, got, results[0])
 		}
-	}
-}
-
-// TestCleanupReleasesUnsandboxedScratchLease is the leak regression from the
-// task-1 review: off/unsandboxed is the DEFAULT sandbox mode, and
-// session_lifecycle.go calls Cleanup() on every session close, so if Cleanup
-// never released the unsandboxedScratch lease, a long-running `evener serve`
-// daemon would hold one open file descriptor (the lease flock) per session
-// for the rest of the process's life — sweepCrashedSessionScratch can only
-// reclaim a lease that is currently acquirable. Verified at the OS level: a
-// fresh flock attempt on the SAME lease file must succeed immediately after
-// Cleanup(), proving the held lock was actually released, not merely that no
-// error was returned.
-func TestCleanupReleasesUnsandboxedScratchLease(t *testing.T) {
-	if runtime.GOOS == "windows" {
-		t.Skip("flock-based lease verification is unix-only")
-	}
-	worktree := t.TempDir()
-	env := NewLocalExecutionEnvironment(worktree)
-	scratch := env.unsandboxedScratchDir()
-	if scratch == "" {
-		t.Fatal("unsandboxedScratchDir provisioning failed")
-	}
-
-	env.Cleanup()
-
-	if scratchLeaseHeld(t, scratch) {
-		t.Fatal("lease still held after Cleanup (leak)")
 	}
 }
